move initiaization to App Router
