+++ apps/examples/nextjs/app/auth/[...nextauth]/route.ts
@@ -1,0 +1,4 @@
+import NextAuth from "next-auth/next"
+import config from "auth"
+const handler = NextAuth(config)
+export { handler as GET, handler as POST }
