+++ apps/examples/nextjs/auth.ts
@@ -1,0 +1,89 @@
+import type {
+  GetServerSidePropsContext,
+  NextApiRequest,
+  NextApiResponse,
+} from "next"
+import type { NextAuthOptions } from "next-auth"
+import { getServerSession } from "next-auth"
+
+import GoogleProvider from "next-auth/providers/google"
+import FacebookProvider from "next-auth/providers/facebook"
+import GithubProvider from "next-auth/providers/github"
+import TwitterProvider from "next-auth/providers/twitter"
+import Auth0Provider from "next-auth/providers/auth0"
+
+// Read more at: https://next-auth.js.org/getting-started/typescript#module-augmentation
+declare module "next-auth/jwt" {
+  interface JWT {
+    /** The user's role. */
+    userRole?: "admin"
+  }
+}
+
+const config = {
+  // https://next-auth.js.org/configuration/providers/oauth
+  providers: [
+    Auth0Provider({
+      clientId: process.env.AUTH0_ID,
+      clientSecret: process.env.AUTH0_SECRET,
+      issuer: process.env.AUTH0_ISSUER,
+    }),
+    FacebookProvider({
+      clientId: process.env.FACEBOOK_ID,
+      clientSecret: process.env.FACEBOOK_SECRET,
+    }),
+    GithubProvider({
+      clientId: process.env.GITHUB_ID,
+      clientSecret: process.env.GITHUB_SECRET,
+    }),
+    GoogleProvider({
+      clientId: process.env.GOOGLE_ID,
+      clientSecret: process.env.GOOGLE_SECRET,
+    }),
+    TwitterProvider({
+      clientId: process.env.TWITTER_ID,
+      clientSecret: process.env.TWITTER_SECRET,
+      version: "2.0",
+    }),
+  ],
+  callbacks: {
+    async jwt({ token }) {
+      token.userRole = "admin"
+      return token
+    },
+  },
+} satisfies NextAuthOptions
+
+export default config
+
+// Helper function to get session without passing config every time
+// https://next-auth.js.org/configuration/nextjs#getserversession
+export function auth(
+  ...args:
+    | [GetServerSidePropsContext["req"], GetServerSidePropsContext["res"]]
+    | [NextApiRequest, NextApiResponse]
+    | []
+) {
+  return getServerSession(...args, config)
+}
+
+// We recommend doing your own environment variable validation
+declare global {
+  namespace NodeJS {
+    export interface ProcessEnv {
+      NEXTAUTH_SECRET: string
+      GITHUB_ID: string
+      GITHUB_SECRET: string
+      FACEBOOK_ID: string
+      FACEBOOK_SECRET: string
+      TWITTER_ID: string
+      TWITTER_SECRET: string
+      GOOGLE_ID: string
+      GOOGLE_SECRET: string
+      AUTH0_ID: string
+      AUTH0_SECRET: string
+      DESCOPE_ID: string
+      DESCOPE_SECRET: string
+    }
+  }
+}
