--- apps/examples/nextjs/next-auth.d.ts
@@ -1,10 +1,0 @@
-import "next-auth/jwt"
-
-// Read more at: https://next-auth.js.org/getting-started/typescript#module-augmentation
-
-declare module "next-auth/jwt" {
-  interface JWT {
-    /** The user's role. */
-    userRole?: "admin"
-  }
-}
