+++ apps/examples/nextjs/pages/api/examples/session.ts
@@ -1,13 +1,12 @@
 // This is an example of how to access a session from an API route
-import { getServerSession } from "next-auth"
-import { authOptions } from "../auth/[...nextauth]"
 
+import { auth } from "auth"
 import type { NextApiRequest, NextApiResponse } from "next"
 
 export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
-  const session = await getServerSession(req, res, authOptions)
+  const session = await auth(req, res)
   res.send(JSON.stringify(session, null, 2))
 }
